+++ src/common/utils/errorHandling/handleCommonErrorResponse.js
@@ -1,0 +1,9 @@
+import { ResponseErrorsContainer } from '@folio/stripes-acq-components';
+
+import { genericErrorStrategy } from './strategies';
+
+export const handleCommonErrorResponse = async ({ response }, params) => {
+  const { handler } = await ResponseErrorsContainer.create(response);
+
+  return handler.handle(genericErrorStrategy(params));
+};