+++ src/common/utils/errorHandling/strategies/index.js
@@ -1,0 +1,1 @@
+export { genericErrorStrategy } from './genericErrorStrategy';